refactor(admin): move operational stores to runes
